2025-2026/Zefei Xie/backend/requirements.txt
numpy==2.3.4
scikit-learn==1.7.2
openai==2.3.0
python-dotenv==1.1.1
transformers==4.57.1
requests==2.32.5
fastapi==0.119.0
pydantic==2.12.2
torch==2.5.1
pillow==12.0.0
tqdm==4.67.1
httpx==0.28.1
sentence-transformers==5.1.2
langchain-openai==0.3.35
langchain-core==0.3.79
langgraph==0.6.10
aiohttp==3.13.0
umap-learn==0.5.9.post2
pydantic-settings==2.11.0
uvicorn==0.37.0
PyYAML==6.0.3
huggingface-hub==0.36.0